🔧 chore: remove unused user API key and model sync files; integrate user API key retrieval into LLMService
